Move helper functions to new file
Once again, staying organized.
